Move getRutils to utils function. Add user test file
