+++ packages/utils/test/utils/client.js
@@ -1,0 +1,13 @@
+import RTokenUtils, { getClient } from '../../src';
+
+export const getRutils = () => {
+  const uri = 'http://localhost:8000/subgraphs/name/rtoken-test';
+  // TODO: set uri based on process.env for local/production testing
+
+  const apolloInstance = getClient({
+    uri,
+    debug: true,
+  });
+  const options = { debug: true };
+  return new RTokenUtils(apolloInstance, options);
+};
